fix(pipeline): keep status summary label on effective model source
The PipelineStatusSummary top bar now resolves the model from any configured source, not just the active tab.

- Move getEffectiveModelSource and EffectiveModelSource into src/lib/modelSource.ts so the status summary does not depend on the VRAM module.

- Update vramEstimate.ts and PipelineStatusSummary.tsx to import from the new module.

- Fix getModelLabel so it no longer short-circuits on the active-tab-only modelSelected boolean.

- Add src/lib/__tests__/modelSource.test.ts with the getEffectiveModelSource cases, including the missing multi-fallback order test.

- Keep hasSelectedModel as the run/export gate for the active source.

- Run pnpm test, pnpm lint, and pnpm build successfully.
